+++ conda_lockfiles/loaders/__init__.py
@@ -1,0 +1,8 @@
+from __future__ import annotations
+
+from .base import BaseLoader
+from .pixi import PixiLoader
+
+__all__ = ["BaseLoader", "PixiLoader", "LOADERS"]
+
+LOADERS = (PixiLoader,)
